Accepts custom EventKeywords with Int64 values when generating the EventSchema (as EventKeywords is a Int64 enum defined in System.Diagnostics.Tracing). Fixes http://entlib.codeplex.com/workitem/33800

+++ source/Src/SemanticLogging/Schema/EventSourceSchemaReader.cs
@@ -81,7 +81,7 @@ namespace Microsoft.Practices.EnterpriseLibrary.SemanticLogging.Schema
 
                 var keywordNames = (string)@event.Attribute("keywords");
 
-                int keywordsMask = 0;
+                long keywordsMask = 0;
                 if (!string.IsNullOrWhiteSpace(keywordNames))
                 {
                     foreach (var keywordName in keywordNames.Split())
@@ -94,7 +94,7 @@ namespace Microsoft.Practices.EnterpriseLibrary.SemanticLogging.Schema
 
                         if (keywordsMaskAtt != null)
                         {
-                            keywordsMask |= Convert.ToInt32(keywordsMaskAtt.Value, 16);
+                            keywordsMask |= Convert.ToInt64(keywordsMaskAtt.Value, 16);
                         }
                     }
                 }
